Fixing a build issue in linux
