+++ npm/pkg/dataplane/ipsets/testutils.go
@@ -1,9 +1,7 @@
 package ipsets
 
 import (
-	"github.com/Azure/azure-container-networking/network/hnswrapper"
 	"github.com/Azure/azure-container-networking/npm/util"
-	"github.com/Microsoft/hcsshim/hcn"
 )
 
 type TestSet struct {
@@ -22,18 +20,6 @@ func CreateTestSet(name string, setType SetType) *TestSet {
 	set.PrefixName = set.Metadata.GetPrefixName()
 	set.HashedName = util.GetHashedName(set.PrefixName)
 	return set
-}
-
-func GetHNSFake() *hnswrapper.Hnsv2wrapperFake {
-	hns := hnswrapper.NewHnsv2wrapperFake()
-	network := &hcn.HostComputeNetwork{
-		Id:   "1234",
-		Name: "azure",
-	}
-
-	hns.CreateNetwork(network)
-
-	return hns
 }
 
 var (
