+++ npm/pkg/dataplane/ipsets/testutils_windows.go
@@ -1,6 +1,22 @@
 package ipsets
 
-import testutils "github.com/Azure/azure-container-networking/test/utils"
+import (
+	"github.com/Azure/azure-container-networking/network/hnswrapper"
+	testutils "github.com/Azure/azure-container-networking/test/utils"
+	"github.com/Microsoft/hcsshim/hcn"
+)
+
+func GetHNSFake() *hnswrapper.Hnsv2wrapperFake {
+	hns := hnswrapper.NewHnsv2wrapperFake()
+	network := &hcn.HostComputeNetwork{
+		Id:   "1234",
+		Name: "azure",
+	}
+
+	hns.CreateNetwork(network)
+
+	return hns
+}
 
 func GetApplyIPSetsTestCalls(_, _ []*IPSetMetadata) []testutils.TestCmd {
 	return []testutils.TestCmd{}
